Var as new temporary type (Issue 23 in GTClang) (#23)
We removed the keyword temporary_storage form gtclang. If users want to allocate temporary storages, var should be used now. This can either be done where storages are defined or in the do-method itself.
If the user is not concerned about the type of literals, var can also be used for that purpose.

Since stencil_functions do not allow for temporary storages, vars are assumed literals in that context.

With this implementation, cleaning up the preprocessor-work for temporaries was also done.