fix to the interval-hash
fixing the interval-hash s.t. it maps equal intervals (in terms of their absoulte bounds) same: 
`Interval{2,2,0,0}` gets the same hash as `Interval{1,3,1,-1}` as they are essentially the same. With this change, comparing two sets of intevals (what apparently checks equality of hashes) yields to the expected result and hence the setcomparison is obsolete