Fixing creation of new intervals for naive codegen (#65)
Some of the splitting expected the handling of open intervals [..) but vertical_regions clearly operate on closed intervals [...].

Also, the constructor for intervals that only gets levels (and therefore uses 2 arguments) breaks once we offset the top-level (as it is the arbitrary number 1<<20). This cleans up all the splitting into proper level / offset constructors taking four arguments.

The changes in the testing had to happen because if you compare elements of a set, c++ applies !(a < b) && !(b < a) rather than a==b. Since these functors are not overloaded in the dawn::interval (as they are somewhat ambiguous for intervals), we use the setcomparison method to force a comparison with the == operator.

Various test were doing the wrong thing as we should handle intervals as closed ones and a lot of tests assumed open ones.