Fix in Codegen for temporaries spanning mulitple multistages (#73)
If temporaries span more than one Multistage (or more than one stencil), they are initialized by the stencil-wrapper. This fix determines the vertical extents that those temporaries must have. The functionality is similar to what was present for naive code generation.

In order to generalize this concept across naive and gridtools code-generation, I slightly adapted the naming convention in the naive code-generartion (going from storage_t XXX_ to storage_t m_XXX. This was the convention used in the gridtools-codegen