Deserialize IIR (#236)
## Technical Description

Provides (de)serialization capabilities to json files. 

## Testing

Was tested with the set of stencils provided in the code gen tests in the sense that code gen from DSL and codegen with an intermediary serialization and de-serialization step produce the same code. Proper testing for (de)serilization and using (de)serialization to test the code gen will be introduced in a separate PR.  

## Dependencies

Depends on [Pull Request 154](https://github.com/MeteoSwiss-APN/gtclang/pull/154) in the gtclang repo