made literal ids strictly negative. adapted test accordingly. removed bad assertion (#268)
Some edge cases in the clang-gridtools test suite were not correctly handled by the changes introduced due to the fix to issue_199 (DoubleSidedMap and general StencilMetaData clean up). AccessID s to literals are now always strictly negative, no matter the context (e.g. even if the literal is in fact a global which was overwritten by a config file). A test was adapted accordingly.