Test deserialization (#281)
## Technical Description 

This introduces testing for the `restoreIIR` function in `OptimizerContext` which computes the derived info of a stencil, given the stencils IIR. This is achieved by constructing a stencil in memory (A) and deserializing an equivalent stencil from disk (B). The in memory IIR, meta data and derived Info of A and B are then compared. 

Since those tests need to access the file system, they are considered integration tests instead of unit tests. These are the first integration tests for dawn. Build files have been adapted accordingly. 

Currently, to ensure equivalent ids between stencils from memory as well as from disk the UUID singleton needs to be reset. This is of course bad practice. The correct way to compare this would be to first try to find a bijective mapping between both id sets. To be discussed if this should go into a different PR are handled herein.