Stage location type [WIP] (#318)
## Technical Description

Add a location type to the stage as well as the reduce over neighbors expresstion. This allows for code-generation that loops over edges/vertices/triangles as the "domain-loop" and reduction to edges/vertices/triangles as the "targets". As a prerequisite, the location type has also been added to the fields themselves. Currently no deduction or consistency checks of location types of stages / reduceOverNeighbor expressions is done.

### Resolves / Enhances

Resolves GT-80 and GT-81

### Example

See example in dawn-c unittest