Getting started (#406)
This PR adds a tutorial for dawn and gtclang. It compiles and executes the same stencil three times: once starting from a stencil written with the GTClang DSL, once starting by using Python to write SIR, and once handing over SIR to Dawn using C++.

The files are found in gtclang/examples/tutorials.