unstructured k offsets (#446)
## Technical Description

This PR now uses the k-offsets introduced previously. That is, vertical offsets are now used in the code generation of `CXXNaiveIco`. NOTE: This currently breaks `lukaslib` compatibility.

### Resolves / Enhances
Fixes #412 

### Testing

A test was introduced to the deactivated c-tests. Follow the `README` in dawn/prototype to generate and run stencil code that uses the k-offsets. (verticalSum / verticalSumDriver)