Add p_grad_c integration test (#481)
Adds an FV3 stencil to the gtclang tests and sets up the integration test infrastructure for more to come later. To enable this, the cmake functions code_generate_examples and code_compile_examples were altered to add a DIRECTORY argument. The new test is automatically built when GTCLANG_BUILD_GT_CPU_EXAMPLES is enabled.