Fix k-Caching in presence of iteration spaces (#566)
## Technical Description

Glocal Iteration spaces were introduced. This has consequences w.r.t to caching. If a stage with an iteration space caches a field, subsequent reads in another stage may "miss" the cache if this other stage has no iteration space or an iteration space outside the first stages'. In these cases, caching needs to be disabled. 

### Testing

There are various simple tests with two stages introduced which check the correct behavior of the caching in presence of iteration spaces.