Consistency location (#594)
## Technical Description

This PR adds consistency type checking for *Location* types. That is, in the unstructured case, the field locations (cell, edge, vertex) are checked to be consistent.

## Examples

The following pseudo code examples would now be rejected by the compiler
```
cell_field = edge_field;
cell_field + edge_field;
cell_field = reduceCellToEdge(edge_field)
```

## Testing

Since the Compiler is now hardwired to perform these tests there are no new tests for correct IIR or SIR. The test in this regard is simply that all tests keep passing. To test rejection of invalid code, a suite of common mistakes in writing IIR have been collected and are located in `dawn/dawn/test/unit-test/dawn/Support/TestTypeChecker.cpp`

## Resolves / Enhances

Fixes #516

## Notes

A note on placement of the type checker. It was envisioned that type checking would be a member function of the IIR and SIR respectively, with the actual TypeChecker residing in the AST folder. Alas, this plan was foiled since the TypeChecker needs to look up if there is `IIRAccessExprData`associated with a `FieldAccessExpr`. Having in mind that this will (hopefully) not remain the only validation carried out, a new module called "Validator" was added