Split parallelizer and optimizer (#608)
First step of optimizer refactoring is splitting the passes into independent parallelizer and optimizer groups. All the passes in the parallelizer are necessary to generate valid parallel code and could be followed directly by code generation.

This PR also begins to refactor the internals of DawnCompiler, e.g. removing a level of unnecessary indirection in parts.

This PR also disables certain tests that break when doing that, so we're collecting these PRs into another branch that can be merged into master once all tests are re-enabled/refactored and pass.