Globals opt away (#611)
This PR adds an AST integrity check pass to the end of the Dawn compiler passes. The current version contains a single check, that assignment statements do not attempt to assign to constant expressions (literals). A semantic error is thrown when this pattern is found in the AST, such as an attempt to reassign a global variable that has been optimized away.