Code Generation for Sparse Dimensions (#654)
## Technical Description

Sparse Dimensions where introduced in PR #602, but no code generation was performed so far. This PR introduces code generation for sparse dimensions. It was noticed that the requirements for sparse dimensions in the presence of nested reductions are not quite clear, hence this case has been disabled. However, it was also noticed that nested reductions were not tested so far, so two tests with regards to this were added. Both `mylib` and `atlas` interfaces were adapted to include sparse dimensions (essentially rank 3 arrays).

## Testing
* Disabled Compiler tests along with Drivers for simple nested reductions, nested reduction including field access, single reduction using sparse dimensions
* `AtlasIntegrationTestCompareOutput` for simple nested reductions, nested reduction including field access, single reduction using sparse dimensions
* A sparse dimension example starting from SIR is also generated using `dawn4py`. The stencil is built (compiled) but not run. 

## Resolves

Fixes #632