Update Grid Type Checker to include FieldDimensions (#686)
## Technical Description

This PR adapts the Gridtype checker to also ensure that `FielDimension`s are of the correct type, since a unstructured version of them has been introduced recently. 

### Resolves / Enhances

Resolves #634 

### Testing

A new test with inconsistent field dimensions has been added, this time starting from SIR instead of IIR like the other tests 

### Dependencies

This PR is independent.