Generalize Weights (#897)
## Technical Description

Until now, weights (in the reduce over neighbors concept) were restricted to be compile time literals. This PR lifts this restriction somewhat and allows the weights to be any arithmetic combination of `FieldAccessExpr`s into dense fields. Accesses into sparse fields or even more complicated constructs like allowing weights to be the result of `ReduceOverNeighborExpr`s are still prohibited. The reason for this is the current interface to the meshing libraries, which would need significant refactoring to enable these constructs. A new validation pass ensuring that the weights conform to these restrictions was introduced.

### Resolves / Enhances

Fixes #870 

### Testing

There are number of new tests:
- Unit tests to ensure that the new validation pass is working correctly in `/dawn/test/unit-test/dawn/Validator/TestUnstructuredDimensionChecker.cpp`
- `dawn/test/unit-test/dawn/IIR/TestIIRSerializer.cpp` was extended to ensure that the generalized weights are (de-)serialized correctly
- The unstructured integration tests were extended with an example close to the ICON diamond FD stencil, using generalized weights

### Dependencies

This PR is independent