Add Diamond Laplacian as SIR test (#942)
## Technical Description

This PR adds the full diamond Laplaican stencil (including Smagorinsky diffusion) in SIR using dawn4py and fixes a small bug in the `ReduceOverNeighbor` concept. 

### Resolves / Enhances

Fixes  #909 

### Testing

This is not tested except that it is checked that the stencil compiles without issues. This is because the driver code is quite complicated. Further verification is done in [this repo](https://github.com/mroethlin/AtlasUtilities)