Fix Faulty Code Generation for Reductions in Cuda Backend (#995)
## Technical Description

Reductions which are not used in this exact fashion:
```
a = reduce(...)
```
but are part of a binary expression, e.g.
```
a = reduce(...) + b
```
lead to nonsensical code in the cuda code generation. This PR fixes this issue. Now, the Cuda code generation performs two passes per statement, first generating all reductions, saving each result to a temporary. In the second pass, this temporary is used in place of the reduction. 

### Example

```
ddt_vn_adv = ddt_vn_adv + reduce(
            coeff_gradekin*z_ekinh,
            Edge > Cell,
            [-1.0, 1.0],
        )
```

### Testing

Two CodeGen tests have been added for this situation. Correct working of the fix was ensured using the icondusk-e2e repo. 

### Dependencies

This PR is independent