Add Horizontal Subdomains for Unstructured Domains (#1027)
## Technical Description

This PR introduces horizontal subdomains for the unstructured case. The design is quite simple; it simply "hijacks" the first `Interval` of the iteration space for the structured case, but redefines it's semantic. Instead of indices, magic numbers are assigned to the lower and upper level of the Interval, which can then be set at runtime using additionally introduced setters in the generated code. This works because in ICON, subdomains are contiguous _in memory_, and ICON stencils are only ever over a single subdomain, or over subdomains which are adjacent in memory. Consider this picture:

![domains](https://user-images.githubusercontent.com/53755555/94805746-27839780-03ed-11eb-8a81-d6554d174237.png)

A ICON stencil may for example run over lateral boundary, the nudging zone, and the interior, but not over lateral boundary + halo. This entails that the contract for the `Interval` does not need to be adapted for the unstructured case and remains meaningful. 

### ToDo

- [x] namespace issues
- [x] Introduce checker that asserts that second interval of an unstructured iteration space is always `nullopt`?

### Testing

New tests have been added to 
* dawn4py integration test 
* unstructured integration test 

### Dependencies

This PR is independent