Ensure Correct Dimensions in Field Assignments (#1036)
## Technical Description

This PR introduces a dimensional correctness check into the `IntegrityChecker`.  The following admissibility matrix is implemented:

```
✓f3D = f1D
✓f3D = f2D
✓f3D = f3D
⨯f2D = f3D
✓f2D = f2D
⨯f2D = f1D
✓f1D = f1D
⨯f1D = f2D
⨯f1D = f3D
```

It is noted that some of the cases (especially f2D = f3D) may make sense (to select a single slice to copy into f2D), especially in the structured world. Currently, we do not have such a use case for ICON, hence we prohibit these kinds of assignments. Due to potential disparity between structured and unstructured admissibility, the check is currently activated for unstructured IIRs only. 

### Testing

* One of the `TestUnstructuredDimensionChecker` tests had to be moved to `TestIntegrityChecker` since that kind of error is now caught earlier on
* New test ensuring `f2D = f3D` is disallowed in the unstructured context (was silently accepted before, producing dangerous code)

### Dependencies

This PR is independent.