Enabling Globals for Unstructured Backends (#1039)
## Technical Description

This PR enables globals for the unstructured backends. Furthermore, an unreported issue where globals were not propagated from the wrapper class to the stencil class was fixed. Additionally, a bug in the unstructured cuda codegen was fixed when translating stencils that only use dense dimensions.

### Resolves / Enhances

Fixes https://github.com/MeteoSwiss-APN/dawn/issues/1030
Fixes https://github.com/MeteoSwiss-APN/dawn/issues/1028

### Notes

The methods to set and get globals in the cuda backend are on the inner stencils. This will be addressed in [this issue](https://github.com/MeteoSwiss-APN/dawn/issues/1038). Also, a method to communicate globals from FORTRAN will need to be devised (not addressed yet). 

### Testing

New tests in dawn4py and a new unstructured integration test to test the correct operation of the `CXXNaiveIco` backend. `CudaIco` backend tested manually. 

### Dependencies

This PR is independent.