Incorporate Sparse Size Deduction to Improve Generated Interface (#1037)
## Technical Description

So far, the sizes of sparse dimensions were provided using a simple "catalogue" of common sizes in the ICON model. This is an appropriate approach if one were to strictly translate ICON only. But for future proofing, and for "combinatorial" testing using dusk, it is desirable to be able to correctly infer these sizes given the chain. @BenWeber42 came up with a clever algorithm to do just that. 

This algorithm is also used to improve the interface to the generated unstructured cuda code, since the outermost stencil class does not need to be templated with the sparse sizes anymore. 

### Testing

* The algorithm is tested using a catalogue of manually computed sized (powered by Bens crayon skills)
* The new interfaces to the unstructured cuda code can not be tested directly using this repo, but they were manually tested using the icondusk-e2e repo. 

### Dependencies

This PR is independent.