Cuda-ICO Backend Assumed that every API Field is a Stencil Field (#1071)
## Technical Description

The Cuda ICO backend assumed that every API field is also a stencil field. It did this while trying to figure out which fields are being written to in order to not clutter the functions copying fields back to the host. This is now fixed.

### Resolves / Enhances

Fixes #1068 

### Example

Examples like:
```
@stencil
def vert_offset_write(
    a: Field[Edge,K], b: Field[Edge,K], unused: Field[Edge, K]):
    with levels_downward as k:
        a = b
```
should not break at cuda code gen time anymore.

### Testing

Unfortunately can not be tested in dawn because we do not support compiling of generated cuda code in the current testing. Manually tested in icondusk-e2e.

### Dependencies

This PR is independent.