Empty stencils (#1079)
## Technical Description

The _statements_ in stencil should not affect the API of the stencil. This should even be true if the stencil is completely empty, i.e. consists of the `pass` statement only. So far this was not true:

* The ico-naive emitted invalid c++ code
* The cuda backend did not emit valid API functions 

This PR addresses both of these shortcomings (even though the issue only mentions the naive backend) 

### Resolves / Enhances

Fixes #1074 

### Example

```
@stencil
def empty(a: Field[Edge]):
    with levels_downward:
        pass
```

### Testing

I don't think its possible to test this from within dawn? 

### Dependencies

This PR is independent