Fix Scoping Issues In Naive Codegen (#1080)
## Technical Description

In the naive backend, each reduction was wrapped into its own scope. This caused some problems in if statements where a reduction was part of the condition, as well as situations where a temporary field was demoted to a local variable. This PR addresses this by removing the scope. The local linear index is now introduced in the lambda capture, which is made possible by marking the lambda as `mutable`. 

### Resolves / Enhances

Fixes #1065 

### Example

A real life example could be
```
  with levels_downward:
    if sum_over(Cell > Edge, 1) < 4:
        f_x = ... #do something at the boundary
    else:
        f_x = sum_over(Cell > Edge, L)
```

### Testing

A test to the effect of the above was added to the unstructured integration checking. 

### Dependencies

This PR is independent