Fix a Stage Merger Problem on Unstructured Stencils  (#1082)
## Technical Description

When the stage merger encountered a candidate stage not on the same location as the current stage, it would simply proceed to the next stage further above. However, this is only legal if the stage to be skipped is independent of the current stage. 

### Resolves / Enhances

Fixes #1053

### Example

Consider this Example
```
        eout0 = sum_over(Edge > Cell > Edge, ein0[Edge > Cell > Edge])
        cout0 = sum_over(Cell > Edge > Cell, cin0[Cell > Edge > Cell])
        eout1 = sum_over(Edge > Cell, cout0)
```
here, `eout1` depends on `cout0`. We are thus not allowed to merge the last stage into the very first 

Consider on the other hand:
```
        eout0 = sum_over(Edge > Cell > Edge, ein0[Edge > Cell > Edge])
        cout0 = sum_over(Cell > Edge > Cell, cin0[Cell > Edge > Cell])
        eout1 = sum_over(Edge > Cell > Edge, ein1[Edge > Cell > Edge]) 
```
here, the stage in the middle is completely independent of stage number 3. The stage merger is thus allowed to proceed to look for other candidates to merge with further above. Indeed, we expect this to be merged as follows:
```
stage 0:
  eout0 = ...
  eout1 = ...
stage 1:
  cout0 = ...
```

### Testing

Tests reflecting the two examples above have been introduced

### Dependencies

This PR is independent.