Do not Demote Sparse Fields (#1083)
## Technical Description

Since we allowed temporary sparse fields they could potentially be replaced with a temporary. If such a temporary sparse field was scheduled to be demoted however, the demotion failed with `Assertion failed: 'exprStmt' first access of temp field to be demoted (i.e lifetime.Begin) is not an 'ExprStmt'`.  since, naturally (in valid SIR), the first access to the temporary is a `LoopStmt`. We now have the option to either not allow sparse fields to be demoted or relax the assertion in `TemporaryHandling.cpp`. Since currently we can not declare variables with anything other then the most basic primitive data types we decided to **not** demote sparse fields for now, consider 
```std::shared_ptr<iir::Stmt> varDeclStmt = instantiation->getMetaData().declareVar(
      true, varname, Type(BuiltinTypeID::Float), assignmentExpr->getRight(), AccessID);
```
[here](https://github.com/MeteoSwiss-APN/dawn/blob/dfd6bea0a3f5c1710fc18c57355c6189efe9c743/dawn/src/dawn/Optimizer/TemporaryHandling.cpp#L166) with built ins defined [here](https://github.com/MeteoSwiss-APN/dawn/blob/dfd6bea0a3f5c1710fc18c57355c6189efe9c743/dawn/src/dawn/Support/Type.h#L24)

### Resolves / Enhances

Fixes: #1050 

### Example

```
@stencil
def tempSparse(      
    inF: Field[Edge],
    outF: Field[Edge]
):
    sparseF: Field[Edge > Cell > Edge]
    with levels_upward:        
        with sparse[Edge > Cell > Edge]:
            sparseF = 1
        outF = sum_over(Edge > Cell > Edge, inF[Edge>Cell>Edge]*sparseF)
```

(assumes syntax proposed [here](https://github.com/dawn-ico/dusk/issues/61))

In this example, `sparseF` is never accessed offset (with regards to its _dense_ index). Hence, it could be demoted to a sparse field. However, since codegen is not ready for this, we prevent this cleanly by not nominating the sparse field to demotion. 

### Testing

Added a test to `TestPassTemporaryType` representing the example above; ensuring that `sparseF` is not demoted. 

### Dependencies

This PR is independent.