Add Padding to Codegen (#1087)
## Technical Description

The ICON model over-commits memory to simplify the domain decomposition. Since this leads to padding on each horizontal level, and we intend to borrow ICON's field pointers directly, we need to respect this padding on the code generation side. While the iteration is only on the number of actual elements, the field indexing (and temp field allocation) needs to account for these paddings.

### Resolves / Enhances

Fixes #1041 

### Example

`./dawn-codegen my_iir.iir --padding-cells=10 --padding-edges=20 --padding-vertices=30 ` to add padding of 10, 20 and 30 to the cell, edge an vertex fields, respectively.

### Testing

Testing from within dawn is limited. A unstructured integration test was added, that mostly tests the assumptions on the driver side. The CUDA side was tested using a manually crafted example. 

### Dependencies

This PR is independent.