Fix a Bug in the Unstr. Dim Checker w.r.t. Index Fields (#1114)
## Technical Description

During the integration workshop it was noted that there seems to be a problem with dawn when using sparse index fields. Further investigation revealed that, while sparse index fields are fully supported, this was rather a case of too strict dimensions checking. It was required that `a[b] <=> dim(a) == dim(b)`, which is too strict if the index fields are used in reductions. This PR basically relaxes the dimensions checking for index fields to the same rules as for normal fields in reductions, while leaving the `a[b] <=> dim(a) == dim(b)` for point-wise situations. 

### Resolves / Enhances

Fixes #1106 

### Example

We basically want to enable stencils like the following, which is a stripped down version [of a real ICON stencil](https://github.com/dawn-ico/ICON-DSL/pull/21). 

```
@stencil
def a_stencil(
    a: Field[Cell, K],
    b: Field[Edge, K],        
    test_field: IndexField[Cell > Edge, K]
):
  with domain.upward as k:
    a = sum_over(Cell > Edge, b[test_field])      
```

Note that relaxing the type checker in this way also allows stencils like:

```
@stencil
def another_stencil(
    a: Field[Cell, K],
    b: Field[Edge, K],        
    c: IndexField[Cell, K],
):
  with domain.upward as k:
    a = sum_over(Cell > Edge, b[c])
```

this may seem strange at first but could basically be used to achieve things like "for some cells iterate the edges a level above while for others the edges one level below", which, given the general nature of ICON stencils, doesn't seem too far fetched. 

### Testing

* A new `dawn4py` test using the first example
* Two new `UnstructuredDimensionsChecker` tests for both examples above

### Dependencies

This PR is independent.