Fix Reductions within if Statements (#1121)
## Technical Description

Unsafe code was generated for stencils of the form:

```
@stencil
def another_stencil(
    a: Field[Cell, K],
    b: Field[Edge, K],        
    mask: Field[Cell, K],
):
  with domain.upward as k:
    if mask:
      a = sum_over(Cell > Edge, b)
```

In the sense that the `sum_over` was executed even if `mask` wasn't true (just the write back to `a` was "masked out"). This is unsafe w.r.t to memory (e.g. in conjunction with index fields which may hold invalid values if `mask` does not hold, and is also a runtime blunder if the mask is sparse.

This PR fixes that by adapting the two-pass strategy for `ReduceOverNeighborExpr`s.

### Testing

Tested using `ICON-DSL`

### Dependencies

This PR is independent.