add VarAccessExpr to integrity checking (#1144)
## Technical Description

During the stencil integration a bug in the dawn integrity checking was noticed with situations of the form:

```
@stencil
def test_stencil(
    levelmask: Field[K],    b: Field[Edge, K]       
) -> None:
    a: Field[Edge, K]
    with domain.upward as k:
        a = 0.0
        if (levelmask[k]):
            a = b   
```

This (valid) code was rejected by dawn if `--default-opt` was set. In the issue it was theorized that the `levelmask` is to blame, since the error message read " trying to assign verticald field to fulld field!". This turned out not to be true; the error was in fact a missing visitor looking at `VarAccessExpr`. If `--default-opt` is set, the field `a` is demoted to a local scalar variable. Since there was no visitor for this variable, the `IntegrityChecker` was indeed comparing the dimensions of `levelmask` and `b`. This PR resolves this by adding the appropriate visitor.

### Resolves / Enhances

Fixes #1143 

### Example

Example is posted above

### Testing

A new test was added to `TestIntegrityChecker`, exercising the example above

### Dependencies

This PR is independent.