Added per field k_size to control extend of verification on per field basis (#1155)
## Technical Description

This PR addresses the problem that in fused stencils the verification is always done for the full range in k-direction, even if some fields do not exist on all k levels (typically some fields exist from [0, nlev) and some [0, nlevpq)). With this PR one can specify the verification k-range on a by-field basis with optional argument in the fortran interface of the setup method.

### Resolves / Enhances

https://gitlab.dkrz.de/dsl/icon-cscs/-/issues/1

### Example

The followin fusion four dusk files into one:
```
-    call setup_mo_solve_nonhydro_stencil_03(p_patch(jg)%dsl_gpu_trimesh, p_patch(jg)%nlev, streamid)
-    call setup_mo_solve_nonhydro_stencil_04(p_patch(jg)%dsl_gpu_trimesh, p_patch(jg)%nlevp1, streamid)
-    call setup_mo_solve_nonhydro_stencil_05(p_patch(jg)%dsl_gpu_trimesh, p_patch(jg)%nlev, streamid)
-    call setup_mo_solve_nonhydro_stencil_06(p_patch(jg)%dsl_gpu_trimesh, p_patch(jg)%nlev, streamid)
+    call setup_fused_mo_solve_nonhydro_stencils_03_04_05_06(p_patch(jg)%dsl_gpu_trimesh, p_patch(jg)%nlevp1, streamid)
```
results in an out of bounds error during the verification. Because only stencil 04 is properly defined for upper k level nlevp1. The verification then accesses out of bounds for the other fields (for example for stencil_06) causing a GPUAssert to trigger.

### Testing

https://gitlab.dkrz.de/dsl/icon-cscs/-/merge_requests/45

### Dependencies

None